# packages/api-contract/openapi/footalert.v1.yaml
openapi: 3.1.0
info:
  title: FootAlert BFF API
  version: 1.0.0
  description: |
    Public contract consumed by FootAlert clients (mobile, web, desktop).
    This spec intentionally uses broad payload schemas because upstream API-Football
    structures can evolve, while preserving endpoint/query/path contracts.
servers:
  - url: http://localhost:3001
    description: Local BFF server
  - url: https://staging-api.footalert.example.com
    description: Staging BFF server
paths:
...
  /v1/teams/standings:
    get:
      summary: Get league standings for teams endpoint
      operationId: getTeamsStandings
      parameters:
        - $ref: '#/components/parameters/LeagueIdQuery'
        - $ref: '#/components/parameters/SeasonQuery'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}:
    get:
      summary: Get team details
      operationId: getTeamById
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/leagues:
    get:
      summary: Get leagues for a team
      operationId: getTeamLeagues
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/fixtures:
    get:
      summary: Get team fixtures
      operationId: getTeamFixtures
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/OptionalSeasonQuery'
        - $ref: '#/components/parameters/OptionalLeagueIdQuery'
        - $ref: '#/components/parameters/OptionalTimezoneQuery'
        - $ref: '#/components/parameters/OptionalNextQuery'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/next-fixture:
    get:
      summary: Get next fixture for team
      operationId: getTeamNextFixture
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/TimezoneQuery'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/advanced-stats:
    get:
      summary: Get advanced team stats
      operationId: getTeamAdvancedStats
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/LeagueIdQuery'
        - $ref: '#/components/parameters/SeasonQuery'
      responses:
        '200':
          $ref: '#/components/responses/OptionalEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/stats:
    get:
      summary: Get team statistics
      operationId: getTeamStats
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/LeagueIdQuery'
        - $ref: '#/components/parameters/SeasonQuery'
      responses:
        '200':
          $ref: '#/components/responses/OptionalEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/players:
    get:
      summary: Get team players (paged)
      operationId: getTeamPlayers
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/LeagueIdQuery'
        - $ref: '#/components/parameters/SeasonQuery'
        - $ref: '#/components/parameters/OptionalPageQuery'
        - $ref: '#/components/parameters/OptionalLimitQuery'
        - $ref: '#/components/parameters/OptionalCursorQuery'
      responses:
        '200':
          $ref: '#/components/responses/PagedEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/squad:
    get:
      summary: Get team squad
      operationId: getTeamSquad
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/transfers:
    get:
      summary: Get team transfers
      operationId: getTeamTransfers
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
  /v1/teams/{id}/trophies:
    get:
      summary: Get team trophies
      operationId: getTeamTrophies
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/ListEnvelopeResponse'
    x-platform-scope:
      - mobile
      - web
      - desktop
...
components:
  securitySchemes:
    MobileSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[1-9]\\d*$'
...
    TimezoneQuery:
      name: timezone
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    OptionalTimezoneQuery:
      name: timezone
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    LeagueIdQuery:
      name: leagueId
      in: query
      required: true
      schema:
        type: string
        pattern: '^[1-9]\\d*$'
    OptionalLeagueIdQuery:
      name: leagueId
      in: query
      required: false
      schema:
        type: string
        pattern: '^[1-9]\\d*$'
...
    SeasonQuery:
      name: season
      in: query
      required: true
      schema:
        type: integer
        minimum: 2000
        maximum: 2100
    OptionalSeasonQuery:
      name: season
      in: query
      required: false
      schema:
        type: integer
        minimum: 2000
        maximum: 2100
...
    OptionalPageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
    OptionalLimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 10
        maximum: 100
    OptionalCursorQuery:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    OptionalNextQuery:
      name: next
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10
...
  responses:
    ListEnvelopeResponse:
      description: API response envelope with list payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListEnvelope'
    PagedEnvelopeResponse:
      description: API response envelope with list + paging
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PagedEnvelope'
    OptionalEnvelopeResponse:
      description: API response envelope with optional payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OptionalEnvelope'
...
    ProblemResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
  schemas:
...
    FlexibleObject:
      type: object
      additionalProperties: true
    CursorPageInfo:
      type: object
      required: [ hasMore, nextCursor, returnedCount ]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        returnedCount:
          type: integer
          minimum: 0
    ListEnvelope:
      type: object
      required: [ response ]
      properties:
        response:
          type: array
          items:
            $ref: '#/components/schemas/FlexibleObject'
        pageInfo:
          oneOf:
            - $ref: '#/components/schemas/CursorPageInfo'
            - type: 'null'
    PagedEnvelope:
      type: object
      required: [ response ]
      properties:
        response:
          type: array
          items:
            $ref: '#/components/schemas/FlexibleObject'
        paging:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        pageInfo:
          oneOf:
            - $ref: '#/components/schemas/CursorPageInfo'
            - type: 'null'
    OptionalEnvelope:
      type: object
      properties:
        response:
          oneOf:
            - $ref: '#/components/schemas/FlexibleObject'
            - type: array
              items:
                $ref: '#/components/schemas/FlexibleObject'
            - type: number
            - type: string
            - type: boolean
            - type: 'null'
    ProblemResponse:
      type: object
      required: [ error, message ]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: 'null'
...